Finished exercise 4.22

+++ part4/blogs-app/controllers/blogsRouter.js
@@ -1,6 +1,27 @@
 const blogsRouter = require("express").Router();
 const Blog = require("../models/blogModel");
 const User = require("../models/userModel");
+const jwt = require("jsonwebtoken");
+
+// Some token utils
+function getTokenFrom(request) {
+    // Defined as middleware in utils, this is just legacy code
+    const authorization = request.get("authorization");
+    if (authorization && authorization.toLowerCase().startsWith("bearer ")) {
+        return String(authorization).replace("Bearer ", "");
+    }
+
+    return undefined;
+}
+
+function getUserFromToken(token) {
+    // Defined as middleware in utils, this is just legacy code
+    const decodedToken = jwt.verify(token, process.env.SECRET);
+    if (!decodedToken.id) {
+        return undefined;
+    }
+    return decodedToken.id;
+}
 
 
 blogsRouter.get("/", (request, response) => {
@@ -19,15 +40,20 @@ blogsRouter.get("/api/blogs", async (request, response) => {
 
 blogsRouter.post("/api/blogs", async (request, response) => {
     try {
+        // Check auth token (middleware in utils)
+        if (!request.user) {
+            return response.status(401).json({ error: "invalid token" });
+        }
+
         // Load user info
-        const user = await User.findById(request.body.user);
+        const user = await User.findById(request.user);
 
         const thisBlog = new Blog({
             title: request.body.title,
             author: request.body.author,
             url: request.body.url,
             likes: request.body.likes,
-            user: user.id
+            user: request.user //ID already comes from token
         });
         const result = await thisBlog.save();
 
@@ -43,6 +69,14 @@ blogsRouter.post("/api/blogs", async (request, response) => {
 
 blogsRouter.post("/api/blogs/:id", async (request, response) => {
     try {
+        // Check auth token (middleware in utils)
+        if (!request.user) {
+            return response.status(401).json({ error: "invalid token" });
+        }
+
+        //Update user id in body so it reflects the user who's updating
+        request.body.user = request.user;
+
         // new: true returns the updated document
         const result = await Blog.findByIdAndUpdate(request.params.id, request.body, { new: true });
         response.status(200).json(result);
@@ -53,7 +87,20 @@ blogsRouter.post("/api/blogs/:id", async (request, response) => {
 
 blogsRouter.delete("/api/blogs/:id", async (request, response) => {
     try {
+        // TODO: Try removing this kind of verifications, it's supposed to be handled by middleware in utils
+        if (!request.user) {
+            return response.status(401).json({ error: "invalid token" });
+        }
+
+        // Check if the user trying to delete the blog is the one who created it
+        // (If no one created the blog, everyone can delete it, this is because initial test didn't have user)
+        const blogToDelete = await Blog.findById(request.params.id);
+        if (String(blogToDelete.user) !== request.user && blogToDelete.user !== undefined) {
+            return response.status(401).json({ error: "User from token does not match user that created/updated the blog entry" });
+        }
+
         await Blog.findByIdAndDelete(request.params.id);
+
         response.status(204).end();
     } catch (error) {
         response.status(400).json({ error: error.message });
